% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param.R
\name{get_cwl_version}
\alias{get_cwl_version}
\title{Get CWL version}
\usage{
get_cwl_version(x)
}
\arguments{
\item{x}{CWL object}
}
\description{
Get CWL version
}
\examples{
library("tidycwl")
library("magrittr")

system.file("cwl/sbg/workflow/gatk4-wgs.json", package = "tidycwl") \%>\%
  read_cwl_json() \%>\%
  get_cwl_version()

system.file("cwl/sbg/workflow/rnaseq-salmon.json", package = "tidycwl") \%>\%
  read_cwl_json() \%>\%
  get_cwl_version()
}
