% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-graph.R
\name{get_nodes}
\alias{get_nodes}
\title{Get nodes in a CWL workflow into a data frame}
\usage{
get_nodes(inputs, outputs, steps)
}
\arguments{
\item{inputs}{Parsed inputs}

\item{outputs}{Parsed outputs}

\item{steps}{Parsed steps}
}
\value{
Data frame containing node information
}
\description{
Get nodes in a CWL workflow into a data frame
}
\examples{
flow <- system.file("cwl/sbg/workflow/gatk4-wgs.json", package = "tidycwl") \%>\% read_cwl_json()
get_nodes(
  flow \%>\% parse_inputs(),
  flow \%>\% parse_outputs(),
  flow \%>\% parse_steps()
) \%>\% str()
}
