% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-param.R
\name{get_steps_doc}
\alias{get_steps_doc}
\title{Get documentation/description for steps}
\usage{
get_steps_doc(steps)
}
\arguments{
\item{steps}{Steps object parsed by \code{\link{parse_steps}}}
}
\value{
Vector of step documentation/descriptions
}
\description{
Get documentation/description for steps
}
\examples{
# steps represented by a dictionary
system.file("cwl/sbg/workflow/rnaseq-salmon.json", package = "tidycwl") \%>\%
  read_cwl_json() \%>\%
  parse_steps() \%>\%
  get_steps_doc()

# steps represented by a list
system.file("cwl/sbg/workflow/rnaseq-salmon.cwl", package = "tidycwl") \%>\%
  read_cwl_yaml() \%>\%
  parse_steps() \%>\%
  get_steps_doc()
}
