% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param.R
\name{get_steps_version}
\alias{get_steps_version}
\title{Get toolkit version for steps}
\usage{
get_steps_version(steps)
}
\arguments{
\item{steps}{Steps object parsed by \code{\link{parse_steps}}}
}
\description{
Get toolkit version for steps
}
\examples{
library("tidycwl")
library("magrittr")

# steps represented by a dictionary
system.file("cwl/sbg/workflow/rnaseq-salmon.json", package = "tidycwl") \%>\%
  read_cwl_json() \%>\%
  parse_steps() \%>\%
  get_steps_version()

# steps represented by a list
system.file("cwl/sbg/workflow/rnaseq-salmon.cwl", package = "tidycwl") \%>\%
  read_cwl_yaml() \%>\%
  parse_steps() \%>\%
  get_steps_version()
}
