% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_cwl}
\alias{is_cwl}
\title{Is this a CWL object?}
\usage{
is_cwl(x)
}
\arguments{
\item{x}{any object}
}
\value{
Logical. \code{TRUE} if it is a CWL object, \code{FALSE} if not.
}
\description{
Is this a CWL object?
}
\examples{
system.file("cwl/sbg/workflow/gatk4-wgs.json", package = "tidycwl") \%>\%
  read_cwl(format = "json") \%>\%
  is_cwl()
}
