% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_draft2}
\alias{is_draft2}
\title{Is this CWL draft2?}
\usage{
is_draft2(x)
}
\arguments{
\item{x}{CWL object}
}
\description{
Is this CWL draft2?
}
\examples{
library("tidycwl")
library("magrittr")

system.file("cwl/sbg/workflow/gatk4-wgs.json", package = "tidycwl") \%>\%
  read_cwl(format = "json") \%>\%
  is_draft2()

system.file("cwl/sbg/workflow/rnaseq-salmon.json", package = "tidycwl") \%>\%
  read_cwl(format = "json") \%>\%
  is_draft2()
}
