% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tool}
\alias{is_tool}
\title{Is this a CWL command line tool?}
\usage{
is_tool(x)
}
\arguments{
\item{x}{CWL object}
}
\description{
Is this a CWL command line tool?
}
\examples{
library("tidycwl")
library("magrittr")

system.file("cwl/sbg/tool/bwa-mem.json", package = "tidycwl") \%>\%
  read_cwl(format = "json") \%>\%
  is_tool()

system.file("cwl/sbg/workflow/gatk4-wgs.json", package = "tidycwl") \%>\%
  read_cwl(format = "json") \%>\%
  is_tool()
}
