% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_v1.0}
\alias{is_v1.0}
\title{Is this CWL v1.0?}
\usage{
is_v1.0(x)
}
\arguments{
\item{x}{CWL object}
}
\value{
Logical. \code{TRUE} if it is a CWL v1.0 object, \code{FALSE} if not.
}
\description{
Is this CWL v1.0?
}
\examples{
system.file("cwl/sbg/workflow/gatk4-wgs.json", package = "tidycwl") \%>\%
  read_cwl(format = "json") \%>\%
  is_v1.0()

system.file("cwl/sbg/workflow/rnaseq-salmon.json", package = "tidycwl") \%>\%
  read_cwl(format = "json") \%>\%
  is_v1.0()
}
