% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_workflow}
\alias{is_workflow}
\title{Is this a CWL workflow?}
\usage{
is_workflow(x)
}
\arguments{
\item{x}{CWL object}
}
\value{
Logical. \code{TRUE} if it is a CWL workflow
(instead of a command line tool), \code{FALSE} if not.
}
\description{
Is this a CWL workflow?
}
\examples{
system.file("cwl/sbg/workflow/gatk4-wgs.json", package = "tidycwl") \%>\%
  read_cwl(format = "json") \%>\%
  is_workflow()

system.file("cwl/sbg/tool/bwa-mem.json", package = "tidycwl") \%>\%
  read_cwl(format = "json") \%>\%
  is_workflow()
}
