% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_meta}
\alias{parse_meta}
\title{Parse the metadata in the CWL workflow}
\usage{
parse_meta(x)
}
\arguments{
\item{x}{CWL object}
}
\value{
List of CWL metadata
}
\description{
Parse the metadata in the CWL workflow
}
\examples{
system.file("cwl/sbg/workflow/gatk4-wgs.json", package = "tidycwl") \%>\%
  read_cwl(format = "json") \%>\%
  parse_meta()
}
