% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_type}
\alias{parse_type}
\title{Parse CWL content type}
\usage{
parse_type(x)
}
\arguments{
\item{x}{CWL object}
}
\description{
Parse CWL content type (Workflow or CommandLineTool)
}
\examples{
library("tidycwl")
library("magrittr")

system.file("cwl/sbg/workflow/gatk4-wgs.json", package = "tidycwl") \%>\%
  read_cwl(format = "json") \%>\%
  parse_type()

system.file("cwl/sbg/tool/bwa-mem.json", package = "tidycwl") \%>\%
  read_cwl(format = "json") \%>\%
  parse_type()
}
