% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cwl}
\alias{print.cwl}
\title{Print CWL objects}
\usage{
\method{print}{cwl}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cwl}.}

\item{...}{Additional parameters for \code{\link{print}} (not used).}
}
\value{
The input \code{cwl} object.
}
\description{
Print a brief summary of the CWL object.
}
\examples{
path <- system.file("cwl/sbg/workflow/rnaseq-salmon.json", package = "tidycwl")
flow <- read_cwl(path, format = "json")
flow
}
