% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycwl-package.R
\docType{package}
\name{tidycwl-package}
\alias{tidycwl}
\alias{tidycwl-package}
\title{tidycwl: Tidy Common Workflow Language Tools and Workflows}
\description{
The Common Workflow Language <https://www.commonwl.org/> is an
    open standard for describing data analysis workflows. This package takes
    the raw Common Workflow Language workflows encoded in JSON or 'YAML'
    and turns the workflow elements into tidy data frames or lists.
    A graph representation for the workflow can be constructed and visualized
    with the parsed workflow inputs, outputs, and steps. Users can embed the
    visualizations in their 'Shiny' applications, and export them
    as HTML files or static images.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sbg.github.io/tidycwl/}
  \item \url{https://github.com/sbg/tidycwl}
  \item Report bugs at \url{https://github.com/sbg/tidycwl/issues}
}

}
\author{
\strong{Maintainer}: Soner Koc \email{soner.koc@sevenbridges.com}

Authors:
\itemize{
  \item Nan Xiao
  \item Dennis Dean \email{dennis.dean@sevenbridges.com} (\href{https://orcid.org/0000-0002-7621-9717}{ORCID})
}

Other contributors:
\itemize{
  \item Seven Bridges Genomics [copyright holder, funder]
}

}
\keyword{internal}
