% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-param.R
\name{get_outputs_id}
\alias{get_outputs_id}
\title{Get ID for outputs}
\usage{
get_outputs_id(outputs)
}
\arguments{
\item{outputs}{Parsed outputs}
}
\value{
Vector of output IDs
}
\description{
Get ID for outputs
}
\examples{
# inputs represented by a dictionary
system.file("cwl/sbg/workflow/rnaseq-salmon.json", package = "tidycwl") \%>\%
  read_cwl_json() \%>\%
  parse_outputs() \%>\%
  get_outputs_id()

# inputs represented by a list
system.file("cwl/sbg/workflow/rnaseq-salmon.cwl", package = "tidycwl") \%>\%
  read_cwl_yaml() \%>\%
  parse_outputs() \%>\%
  get_outputs_id()
}
