% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-param.R
\name{get_outputs_label}
\alias{get_outputs_label}
\title{Get label for outputs}
\usage{
get_outputs_label(outputs)
}
\arguments{
\item{outputs}{Parsed outputs}
}
\value{
Vector of output labels
}
\description{
Get label for outputs
}
\examples{
# inputs represented by a dictionary
system.file("cwl/sbg/workflow/rnaseq-salmon.json", package = "tidycwl") \%>\%
  read_cwl_json() \%>\%
  parse_outputs() \%>\%
  get_outputs_label()

# inputs represented by a list
system.file("cwl/sbg/workflow/rnaseq-salmon.cwl", package = "tidycwl") \%>\%
  read_cwl_yaml() \%>\%
  parse_outputs() \%>\%
  get_outputs_label()
}
