% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param.R
\name{get_steps_revision}
\alias{get_steps_revision}
\title{Get revision number for steps}
\usage{
get_steps_revision(steps)
}
\arguments{
\item{steps}{Steps object parsed by \code{\link{parse_steps}}}
}
\description{
Get revision number for steps
}
\examples{
library("tidycwl")
library("magrittr")

# steps represented by a dictionary
system.file("cwl/sbg/workflow/rnaseq-salmon.json", package = "tidycwl") \%>\%
  read_cwl_json() \%>\%
  parse_steps() \%>\%
  get_steps_revision()

# steps represented by a list
system.file("cwl/sbg/workflow/rnaseq-salmon.cwl", package = "tidycwl") \%>\%
  read_cwl_yaml() \%>\%
  parse_steps() \%>\%
  get_steps_revision()
}
