% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_workflow}
\alias{is_workflow}
\title{Is this a CWL workflow?}
\usage{
is_workflow(x)
}
\arguments{
\item{x}{CWL object}
}
\description{
Is this a CWL workflow?
}
\examples{
library("tidycwl")
library("magrittr")

system.file("cwl/sbg/workflow/gatk4-wgs.json", package = "tidycwl") \%>\%
  read_cwl(format = "json") \%>\%
  is_workflow()

system.file("cwl/sbg/tool/bwa-mem.json", package = "tidycwl") \%>\%
  read_cwl(format = "json") \%>\%
  is_workflow()
}
