% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_cwl_json}
\alias{read_cwl_json}
\title{Read a CWL file (JSON format) into a list}
\usage{
read_cwl_json(file)
}
\arguments{
\item{file}{A file path, JSON string, or connection.}
}
\value{
List representation of the input CWL
}
\description{
Read a CWL file (JSON format) into a list
}
\examples{
system.file("cwl/sbg/workflow/rnaseq-salmon.json", package = "tidycwl") \%>\%
  read_cwl_json()
}
