% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_cwl_yaml}
\alias{read_cwl_yaml}
\title{Read a CWL file (YAML format) into a list}
\usage{
read_cwl_yaml(file)
}
\arguments{
\item{file}{A file path, YAML string, or connection.}
}
\description{
Read a CWL file (YAML format) into a list
}
\examples{
library("tidycwl")
library("magrittr")

system.file("cwl/sbg/workflow/rnaseq-salmon.cwl", package = "tidycwl") \%>\%
  read_cwl_yaml()
}
