% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycwl-package.R
\docType{package}
\name{tidycwl-package}
\alias{tidycwl}
\alias{tidycwl-package}
\title{tidycwl: Tidy Common Workflow Language Tools and Workflows}
\description{
Read, parse, and render Common Workflow Language tools and workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sbg.github.io/tidycwl/}
  \item \url{https://github.com/sbg/tidycwl}
  \item Report bugs at \url{https://github.com/sbg/tidycwl/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{nan.xiao@sevenbridges.com} (0000-0002-0250-5673)

Authors:
\itemize{
  \item Soner Koc \email{soner.koc@sevenbridges.com}
  \item Dennis Dean \email{dennis.dean@sevenbridges.com}
}

Other contributors:
\itemize{
  \item Seven Bridges Genomics [copyright holder, funder]
}

}
\keyword{internal}
